Refactor ecs.h
