+++ include/pixbench/entity.h
@@ -1,6 +1,8 @@
 #ifndef ENTITY_HEADER
 #define ENTITY_HEADER
 
+#include "pixbench/engine_config.h"
+#include <bitset>
 #include <cstddef>
 
 
@@ -21,6 +23,14 @@ struct EntityID {
     bool operator!=(EntityID const& other) const {
         return (this->id != other.id) || (this->version != other.version);
     }
+};
+
+
+struct EntityInfo {
+    EntityID entityid;                              //!< Unique entity's identifier
+    bool active = false;                            //!< active status, true means this entity is active
+    size_t current_version = 0;                     //!< version, to prevent old disabled id to be used by user
+    std::bitset<MAX_COMPONENTS> component_mask;     //!< mask of assigned components to this entity
 };
