+++ pixbench/game.cpp
@@ -1,5 +1,6 @@
 #include "pixbench/game.h"
 #include "pixbench/ecs.h"
+#include "pixbench/systems.h"
 #include "pixbench/engine_config.h"
 #include "pixbench/utils.h"
 #include "pixbench/vector2.h"
